Fix for a major bug that caused hanging when a large number of commands were executed.
This fix also reduces the memory footprint of the shell since it imposes the max number of commands that are allowed in the list of commands executed. Once this max is hit a cleanup is performed at which time the list of commands being executed is reduced to half of the max amount.

This is now RootTools 3.1